% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kniterate.R
\name{kniterate}
\alias{kniterate}
\title{Given the root of a package, knit .Rmd files into R code.}
\usage{
kniterate(package_root = getwd())
}
\description{
Given the root of a package, knit .Rmd files into R code.
}
